+++ src/events/Client/ready.ts
@@ -4,6 +4,6 @@ export default createEvent<"ready">({
     name: "ready",
     once: true
 }, async (client) => {
-    client.logger.info("event(ready)", `${client.user.username} is connected to Discord API & Ready!`);
+    client.logger.info("event(ready)", `${client.user.username} is ready!`);
     await deployApplicationCommands(client)
 })